Integrated pLabel support into GUI
