+++ src/croco_wx.py
@@ -17,7 +17,7 @@ import wx.adv
 import croco
 from pandas import read_csv
 
-class CrocoMainWindow(wx.Frame):
+class CroCoMainFrame(wx.Frame):
 
     def __init__(self):
         # ensure the parent's __init__ is called
@@ -31,27 +31,42 @@ class CrocoMainWindow(wx.Frame):
         self.panel = wx.Panel(self)
 
         # define the croco read and write options and link to the modules
-        self.availReads = {'pLink1': croco.pLink1.Read,
-                           'pLink2': croco.pLink2.Read,
-                           'Kojak': croco.Kojak.Read,
-                           'Kojak + Percolator': croco.KojakPercolator.Read,
-                           'StavroX': croco.StavroX.Read,
-                           'Xi': croco.Xi.Read,
-                           'Xi + XiFDR': croco.XiSearchFDR.Read,
-                           'xQuest': croco.xQuest.Read,
-                           'xTable': croco.xTable.Read}
-        self.availWrites =  {'xTable': croco.xTable.Write,
-                             'xVis': croco.xVis.Write,
-                             'xiNet': croco.xiNET.Write,
-                             'DynamXL': croco.DynamXL.Write,
-                             'xWalk': croco.xWalk.Write}
+        # each value of a dict is a list of [options, functionToCall]
+        self.availReads = {'pLink1': [croco.pLink1.Read, []],
+                           'pLink2': [croco.pLink2.Read, []],
+                           'Kojak': [croco.Kojak.Read, [('Kojak rawfile',
+                                                         'file',
+                                                         'Please provide the name of the rawfile for the data')]],
+                           'Kojak + Percolator': [croco.KojakPercolator.Read, []],
+                           'StavroX': [croco.StavroX.Read, [('SSF file',
+                                                             'file',
+                                                             'Please provide an SSF file as returned by StavroX')]],
+                           'Xi': [croco.Xi.Read, []],
+                           'Xi + XiFDR': [croco.XiSearchFDR.Read,[('Xi search file',
+                                                                   'file',
+                                                                   'Please provide Xi output file')]],
+                           'xQuest': [croco.xQuest.Read, []],
+                           'xTable': [croco.xTable.Read, []]}
 
-        # Additional information required for some of the conversions
-        self.reqInput = {'Kojak': ['Kojak rawfile',],
-                         'StavroX': ['SSF file',],
-                         'Xi + FDR': ['Xi search file'],
-                         'xWalk': ['PDB to map xlinks to', 'PDB Atom code'],
-                         }
+        self.availWrites =  {'xTable': [croco.xTable.Write, []],
+                             'xVis': [croco.xVis.Write, []],
+                             'xiNet': [croco.xiNET.Write, []],
+                             'DynamXL': [croco.DynamXL.Write, []],
+                             'xWalk': [croco.xWalk.Write, [('PDB to map xlinks to',
+                                                            'file',
+                                                            'Please provide a PDB file you want to map the xlinks to.\n'+\
+                                                            'No special format is required as only the filename is used'),
+                                                           ('PDB Atom code',
+                                                            'input',
+                                                            'Provide an PDB atom code (e.g. CB) for distance calculation')]],
+                             'pLabel': [croco.pLabel.Write, [('Dir containing mgf files',
+                                                              'dir',
+                                                              'Provide path to a fodler containing mgf-files corresponding to the'+\
+                                                              'rawfile names in the input file'),
+                                                             ('Xlinker as referenced by pLabel',
+                                                              'input',
+                                                              'Provide a cross-linker name (e.g. BS3) that is used by pLabel'+\
+                                                              'to calculate the potentially linked amino acids')]]}
 
         # set triggers allowing the start button to unhide
         self.readSet = False
@@ -61,12 +76,16 @@ class CrocoMainWindow(wx.Frame):
         self.makeMenuBar()
 
         # and a status bar
-        self.CreateStatusBar()
-        self.SetStatusText("Welcome to CroCo!")
+        #self.CreateStatusBar()
+        #self.SetStatusText("Welcome to CroCo!")
 
         # load widgets
         # ALWAYS AFTER LOADING MENU AND STATUS BAR!!!!
         self.createWidgets()
+
+        # arguments to pass at function call
+        self.inputArgs = ''
+        self.outputArgs = ''
 
     def createWidgets(self):
 
@@ -77,30 +96,42 @@ class CrocoMainWindow(wx.Frame):
         input_lbl = wx.StaticText(self.panel,wx.ID_ANY, label='Input', style=wx.ALIGN_CENTER )
         self.inputButton = wx.Button(self.panel, label='Load file(s)')
         self.inputButton.Enable(False)
-        self.readFormat = wx.Choice(self.panel, choices=list(self.availReads.keys()))
+        self.readFormat = wx.Choice(self.panel, choices=sorted(list(self.availReads.keys())))
 
         output_lbl = wx.StaticText(self.panel,wx.ID_ANY, label='Output', style=wx.ALIGN_CENTER )
         self.outputButton = wx.Button(self.panel, label='Write to')
         self.outputButton.Enable(False)
-        self.writeFormat = wx.Choice(self.panel, choices=list(self.availWrites.keys()))
+        self.writeFormat = wx.Choice(self.panel, choices=sorted(list(self.availWrites.keys())))
 
         self.controlStart = wx.Button(self.panel, label='Start')
         self.controlStart.Enable(False)
         controlQuit = wx.Button(self.panel, label='Quit')
 
-        gauge = wx.Gauge(self.panel, range=100)
+        helpButton = wx.ContextHelpButton(self.panel)
+
+        # progress bar
+        # gauge = wx.Gauge(self.panel, range=100)
 
         ## create Connects
 
         self.readFormat.Bind(wx.EVT_CHOICE, self.OnReadFormat)
+        self.readFormat.Bind(wx.EVT_HELP,
+                             lambda evt: self.Info('Select data-format for input', caption='Help'))
         self.writeFormat.Bind(wx.EVT_CHOICE, self.OnWriteFormat)
+        self.writeFormat.Bind(wx.EVT_HELP,
+                              lambda evt: self.Info('Select data-format for output', caption='Help'))
 
         self.inputButton.Bind(wx.EVT_BUTTON, self.OnOpenSwitch)
+        self.inputButton.Bind(wx.EVT_HELP,
+                              lambda evt: self.Info('Opens a dialog to select file for input', caption='Help'))
         self.outputButton.Bind(wx.EVT_BUTTON, self.OnOutputDir)
-
+        self.outputButton.Bind(wx.EVT_HELP,
+                              lambda evt: self.Info('Opens a dialog to select output dir', caption='Help'))
+        
         controlQuit.Bind(wx.EVT_BUTTON, self.OnExit)
         self.controlStart.Bind(wx.EVT_BUTTON, self.OnStart)
-
+        self.controlStart.Bind(wx.EVT_HELP,
+                               lambda evt: self.Info('Start the conversion', caption='Help'))
         ## define the layout
         # define the sizers (main widget layouts) used in the app
         topSizer = wx.BoxSizer(wx.VERTICAL)
@@ -124,13 +155,14 @@ class CrocoMainWindow(wx.Frame):
         # Set start and quit buttons
         controlSizer.Add(self.controlStart, 1, wx.RIGHT, 5)
         controlSizer.Add(controlQuit, 0, wx.LEFT | wx.ALIGN_RIGHT, 5)
+        controlSizer.Add(helpButton, 0, wx.LEFT | wx.ALL | wx.EXPAND, 5)
 
         # assign lower sizers to the top-level sizer
         topSizer.Add(loadSizerContainer, 0, wx.ALL|wx.EXPAND, 5)
         # topSizer.Add(reviewSizer, 0, wx.ALL|wx.EXPAND, 5)
         topSizer.Add(wx.StaticLine(self.panel), 0, wx.ALL|wx.EXPAND, 5)
-        topSizer.Add(controlSizer, 0, wx.ALL|wx.EXPAND, 5)
-        topSizer.Add(gauge, 0,wx.ALL|wx.EXPAND, 5)
+        topSizer.Add(controlSizer, 0, wx.EXPAND|wx.RIGHT, 5)
+        # topSizer.Add(gauge, 0,wx.ALL|wx.EXPAND, 5)
 
         # assign the top sizer to the panel i.e. main layout instance
         self.panel.SetSizer(topSizer)
@@ -263,7 +295,7 @@ class CrocoMainWindow(wx.Frame):
 
         aboutInfo = wx.adv.AboutDialogInfo()
         aboutInfo.SetName("The CroCo cross-link converter")
-        aboutInfo.SetVersion('0.3')
+        aboutInfo.SetVersion('0.4')
         aboutInfo.SetDescription("Graphical interface to convert results from "+\
                                  "data analysis of chemical cross-linking "+\
                                  "mass-spectrometry experiments.")
@@ -290,20 +322,30 @@ class CrocoMainWindow(wx.Frame):
         Show dialog if additional user input is required.
         Otherwise start the conversion
         """
-        if (self.theReadFormat in self.reqInput.keys()) | \
-            (self.theWriteFormat in self.reqInput.keys()):
-                
-                dlg = wx.MessageDialog(parent=None,
-                                       message="Not yet implemented!", 
-                                       caption="Information",
-                                       style=wx.OK|wx.ICON_INFORMATION)
-                dlg.ShowModal()
-                dlg.Destroy()
+        # Check if there are options to ask for
+        if len(self.availReads[self.theReadFormat][1]) > 0 or\
+            len(self.availWrites[self.theWriteFormat][1]) > 0:
+
+            # init the OptionsWindow as child
+            OptionsFrame = CroCoOptionsFrame(self)
+
+            # set the variables in the child window
+            OptionsFrame.InputOptionsToAsk = self.availReads[self.theReadFormat][1]
+            OptionsFrame.OutputOptionsToAsk = self.availWrites[self.theWriteFormat][1]
+
+            # update the controls according to the variables
+            OptionsFrame.updateOptions()
+            # show the window
+            OptionsFrame.Show()
 
         else:
-            self.OnRun(event)
+            # reset args lists
+            self.inputArgs = ''
+            self.outputArgs = ''
+            
+            self.onRun(event)
 
-    def OnRun(self, event):
+    def onRun(self, event):
         """
         Collect all necessary information from self and start the
         conversion by calling the actual conversion script
@@ -317,7 +359,12 @@ class CrocoMainWindow(wx.Frame):
 
         for f in self.theInput:
             try:
-                xtable = self.availReads[self.theReadFormat](f)
+                if len(self.inputArgs) > 0 :
+                    print('Found input args: "{}"'.format(self.inputArgs))
+                    xtable = self.availReads[self.theReadFormat][0](f, *self.inputArgs)
+                else:
+                    print('Using standard args for input')
+                    xtable = self.availReads[self.theReadFormat][0](f)  
                 print('{}: Table succesfully read!'.format(f))
             except Exception as e:
                 self.Warning(str(e))
@@ -335,11 +382,17 @@ class CrocoMainWindow(wx.Frame):
 
             try:
                 print('Writing table in {} format to {}'.format(self.theWriteFormat, outpath))
-                self.availWrites[self.theWriteFormat](xtable, outpath)
+                if len(self.outputArgs) > 0:
+                    print('Found output args: "{}"'.format(self.outputArgs))
+                    self.availWrites[self.theWriteFormat][0](xtable, outpath, *self.outputArgs)
+                else:
+                    print('Using standard args for output')
+                    self.availWrites[self.theWriteFormat][0](xtable, outpath)
                 print('Table successfully written!')
             except Exception as e:
                 self.Warning('Conversion of {} was '.format(f) +
                                    'not successfull:{}'.format(str(e)))
+                print(e)
                 was_error = True
                 break
 
@@ -348,10 +401,204 @@ class CrocoMainWindow(wx.Frame):
                      'File(s) successfully written ' +
                      'to {}!'.format(outpath))
 
+class CroCoOptionsFrame(wx.Frame):
+    """
+    This is a Child Frame to CroCoMainWindow asking the user for input
+    regarding the possible optinos of a submodule
+    """
+
+    def __init__(self, parent):
+        wx.Frame.__init__(self,
+                          None,
+                          wx.ID_ANY,
+                          style= wx.SYSTEM_MENU | wx.CAPTION | wx.CLOSE_BOX,
+                          title='Options')
+
+        self.panel = wx.Panel(self)
+
+        # generate a variable containing the parent to allow pushing back
+        # the user-provided input
+        self.parent = parent
+
+        # in these variables the options from CroCoMainWindow will be written
+        self.InputOptionsToAsk = ''
+        self.OutputOptionsToAsk = ''
+
+    def updateOptions(self):
+
+        def CreateOptionsContainer(self, option, listofOptions):
+            """
+            Args:
+                option: tuple with (label, type of input)
+                listOfOptions: a list object to append to user-input to
+            Returns:
+                optionContainer: wx.BoxSizer object
+            """
+
+            optionLabel = wx.StaticText(self.panel,
+                                        wx.ID_ANY,
+                                        label=option[0],
+                                        style=wx.ALIGN_CENTER)
+            optionLabel.Bind(wx.EVT_HELP,
+                             lambda evt: self.Info(option[2], caption='Help'))
+
+            if option[1] == 'file':
+                optionButton = wx.Button(self.panel, label='Load file', name=option[0])
+                # Bind using a lambda function to be able to pass multiple args
+                # to the called function
+                # see: https://wiki.wxpython.org/Passing%20Arguments%20to%20Callbacks
+                optionButton.Bind(wx.EVT_BUTTON,
+                                  lambda evt, appendTo=listofOptions, label=option[0]: self.onOpenFile(evt, appendTo, label))
+                optionButton.Bind(wx.EVT_HELP,
+                                  lambda evt: self.Info(option[2], caption='Help'))
+                        
+            elif option[1] == 'dir':
+                optionButton = wx.Button(self.panel, label='Open directory', name=option[0])
+                optionButton.Bind(wx.EVT_BUTTON,
+                                  lambda evt, appendTo=listofOptions, label=option[0]: self.onOpenDir(evt, appendTo, label))
+                optionButton.Bind(wx.EVT_HELP,
+                                  lambda evt: self.Info(option[2], caption='Help'))
+            elif option[1] == 'input':
+                optionButton = wx.TextCtrl(self.panel, value="", name=option[0])
+                self.textCtrls[option[0]] = optionButton
+
+            else:
+                raise Exception('Wrong options format for option: {}'\
+                                    .format(option[0]))
+
+            optionContainer = wx.BoxSizer(wx.HORIZONTAL)
+            optionContainer.Add(optionLabel, 0, wx.ALL|wx.EXPAND, 5)
+            optionContainer.Add(optionButton, 0, wx.ALL|wx.EXPAND, 5)
+
+            return optionContainer
+
+        # Dicts mapping the label names to the user-provided input
+        self.inputOptionsToUserInput = {}
+        self.outputOptionsToUserInput = {}
+
+        # dict mapping text labels to input text fields
+        self.textCtrls = {}
+
+        # lists containign the label, button sizers for passing to higher
+        # level sizer
+        InputOptionContainers = []
+        OutputOptionContainers = []
+
+        for option in self.InputOptionsToAsk:
+
+            optionContainer = CreateOptionsContainer(self,
+                                                     option,
+                                                     self.inputOptionsToUserInput)
+            InputOptionContainers.append((optionContainer, 0, wx.ALL|wx.EXPAND, 5))
+
+        for option in self.OutputOptionsToAsk:
+            optionContainer = CreateOptionsContainer(self,
+                                                     option,
+                                                     self.outputOptionsToUserInput)
+            OutputOptionContainers.append((optionContainer, 0, wx.ALL|wx.EXPAND, 5))
+
+        ### createWidgets
+
+        inputSizer = wx.BoxSizer(wx.VERTICAL)
+        inputSizer.AddMany(InputOptionContainers)
+
+        outputSizer = wx.BoxSizer(wx.VERTICAL)
+        outputSizer.AddMany(OutputOptionContainers)
+
+        optionSizer = wx.BoxSizer(wx.HORIZONTAL)
+        optionSizer.Add(inputSizer, 0, wx.ALL|wx.EXPAND, 5)
+        optionSizer.Add(outputSizer, 0, wx.ALL|wx.EXPAND, 5)
+
+        ## Controls
+        okayButton = wx.Button(self.panel, label='Okay')
+        closeButton = wx.Button(self.panel, label='Close')
+        helpButton = wx.ContextHelpButton(self.panel)
+
+        controlSizer = wx.BoxSizer(wx.HORIZONTAL)
+        controlSizer.Add(okayButton, 0, wx.EXPAND|wx.LEFT, 5)
+        controlSizer.Add(closeButton, 0, wx.EXPAND|wx.RIGHT, 5)
+        controlSizer.Add(helpButton, 0, wx.LEFT | wx.ALL | wx.EXPAND, 5)
+
+        ## topSizer
+        topSizer = wx.BoxSizer(wx.VERTICAL)
+        topSizer.Add(optionSizer, 0, wx.ALL|wx.EXPAND, 5)
+        topSizer.Add(wx.StaticLine(self.panel), 0, wx.ALL|wx.EXPAND, 5)
+        topSizer.Add(controlSizer, 0, wx.ALL|wx.EXPAND, 5)
+
+        # assign the top sizer to the panel i.e. main layout instance
+        self.panel.SetSizer(topSizer)
+        topSizer.Fit(self)
+
+        closeButton.Bind(wx.EVT_BUTTON, self.onCancel)
+        okayButton.Bind(wx.EVT_BUTTON, self.onOkay)
+
+    def onCancel(self, event):
+        print('Closing options window')
+        self.Close()
+
+    def onOkay(self, event):
+
+        for label, t in self.InputOptionsToAsk:
+            if t == 'input':
+                self.inputOptionsToUserInput[label] = self.textCtrls[label].GetValue()
+
+        for label, t in self.OutputOptionsToAsk:
+            if t == 'input':
+                self.outputOptionsToUserInput[label] = self.textCtrls[label].GetValue()
+
+        inputArgs = []
+        for label, _ in self.InputOptionsToAsk:
+            inputArgs.append(self.inputOptionsToUserInput[label])
+        outputArgs = []
+        for label, _ in self.OutputOptionsToAsk:
+            outputArgs.append(self.outputOptionsToUserInput[label])  
+
+        print('=== Input ===')
+        for key in self.inputOptionsToUserInput:
+            print('{}: {}'.format(key, self.inputOptionsToUserInput[key]))
+        
+        print('=== Output ===')
+        for key in self.outputOptionsToUserInput:
+            print('{}: {}'.format(key, self.outputOptionsToUserInput[key]))
+        
+        self.parent.inputArgs = inputArgs
+        self.parent.outputArgs = outputArgs
+        
+        self.parent.onRun(event)
+
+        print('Closing options window after passing options to main window')
+        self.Close()
+
+    def onOpenFile(self, event, dictToAppend, label):
+       dlg = wx.FileDialog(self,
+                           message="Choose a file for input",
+                           defaultDir=os.getcwd(),
+                           defaultFile="",
+                           wildcard="*.*",
+                           style=wx.FD_MULTIPLE)
+       if dlg.ShowModal() == wx.ID_OK:
+            dictToAppend[label] = dlg.GetPath()
+       dlg.Destroy()
+
+    def onOpenDir(self, event, dictToAppend, label):
+        dlg = wx.DirDialog(self,
+                           message="Choose one directory for Input:",
+                           defaultPath=os.getcwd(),
+                           style=wx.DD_DEFAULT_STYLE | wx.DD_DIR_MUST_EXIST)
+
+        if dlg.ShowModal() == wx.ID_OK:
+            dictToAppend[label] = dlg.GetPath()
+        dlg.Destroy()
+
+    def Info(self, message, caption = 'CroCo'):
+        dlg = wx.MessageDialog(self, message, caption, wx.OK | wx.ICON_INFORMATION)
+        dlg.ShowModal()
+        dlg.Destroy()
+
 
 if __name__ == '__main__':
     # When this module is run (not imported) then create the app, the
     # frame, show it, and start the event loop.
     app = wx.App()
-    frm = CrocoMainWindow().Show()
+    frm = CroCoMainFrame().Show()
     app.MainLoop()
